all: introduce FlexContainer resource.
FlexContainer resources are flexible containers holding a set of custom
attributes. A FlexContainer has a container definition specifying a
fixed list of custom attribute.
Each FlexContainer specialization is defined by a XSD file located into
commons project.
A FlexContainer resource may be associated with a OSGi
FlexContainerService. A FlexContainerService is responsible to retrieve
and set the custom attribute value of the FlexContainer.

Signed-off-by: BOLLE Sebastien <sebastien.bolle@orange.com>
Signed-off-by: BONNARDEL Gregory <gbonnardel.ext@orange.com>