all: introduce announcement feature
The announcement feature can be used with only application and
FlexContainer.
An application can be announced if and only if it is located under the
CSE Base.
A FlexContainer can be announced if and only if it is located under an
application or a FlexContainer.
The default admin/admin account is used to announce resources.
AnnounceAttributes field is not handled.
A resource can be announced only at the creation (UPDATE requests are
not handled). 

Signed-off-by: BOLLE Sebastien <sebastien.bolle@orange.com>
Signed-off-by: BONNARDEL Gregory <gbonnardel.ext@orange.com>